requirements.txt
black
python-dotenv
ipython
tweepy
